[Coding Dojo: Session 5]: Game of Life
- Date: 12/08/2015
- Problem Description:  The Game of Life (or simply Life) is not a game
in the conventional sense. There are no players, and no winning or
losing. Once the "pieces" are placed in the starting position, the rules
determine everything that happens later. Nevertheless, Life is full of
surprises! In most cases, it is impossible to look at a starting
position (or pattern) and see what will happen in the future.